requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dataclasses-json==0.6.7
deepdiff==8.0.1
emoji==2.14.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
gitpython==3.1.43
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx-sse==0.4.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema-specifications==2024.10.1
jsonschema==4.23.0
langchain-community==0.3.5
langchain-core==0.3.15
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langchain==0.3.7
langdetect==1.0.9
langsmith==0.1.142
lxml==5.3.0
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.13.3
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
orderly-set==5.2.2
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.0.0
pycparser==2.22
pydantic-core==2.23.4
pydantic-settings==2.6.1
pydantic==2.9.2
pydeck==0.9.1
pygments==2.18.0
pypdf2==3.0.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
pyyaml==6.0.2
rapidfuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests-toolbelt==1.0.0
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.35
streamlit==1.40.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.2
triton==3.1.0
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
unstructured-client==0.25.9
unstructured==0.16.5
urllib3==2.2.3
watchdog==5.0.3
webencodings==0.5.1
wrapt==1.16.0
yarl==1.17.1
